+++ .github/workflows/python-app.yml
@@ -56,7 +56,7 @@ jobs:
         pip install -r requirements-dev.txt
     - name: Test with pytest
       run: |
-        python -m pytest tests --cov=src/src_critplot --cov-report=xml --cov-report=term-missing --ignore=tests/critplot/opengl
+        python -m pytest tests --cov=src --cov-report=xml --cov-report=term-missing --ignore=tests/opengl
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
       with:
@@ -87,7 +87,7 @@ jobs:
         pip install -r requirements-dev.txt
     - name: Test with pytest
       run: |
-        python -m pytest tests --ignore=src/core_atomistic_qt --ignore=tests/critplot/opengl
+        python -m pytest tests --ignore=tests/opengl
 
   test-mac:
 
@@ -106,4 +106,4 @@ jobs:
             pip install -r requirements-dev.txt
     - name: Test with pytest
       run: |
-        python -m pytest tests --ignore=src/core_atomistic_qt --ignore=tests/critplot/opengl
+        python -m pytest tests --ignore=tests/opengl
