+++ src/models/cp.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from src_core_atomistic.atom import Atom
+from core_atomistic.atom import Atom
 
 
 class CriticalPoint(Atom):
