+++ src/models/plane.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 import numpy as np
