+++ tests_core_atomistic/test_helpers.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from src_core_atomistic import helpers
+from core_atomistic import helpers
 
 
 def test_helpers_spacedel():
